requirements.txt
adrf==0.1.1
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.5
cloudinary==1.33.0
coreschema==0.0.4
cryptography==41.0.4
Django==4.2.3
django-after-response==0.2.2
django-autoslug==1.9.9
django-cities-light==3.9.2
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
h11==0.14.0
httptools==0.6.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
pluggy==1.2.0
progressbar2==4.2.0
psycopg==3.2.1
psycopg-binary==3.2.1
pycparser==2.21
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-decouple==3.8
python-dotenv==1.0.0
python-utils==3.7.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.0
watchfiles==0.20.0
websockets==11.0.3
whitenoise==6.5.0
